refactor: remove empty interface 'User', rename class 'AbstractUser' to 'User'
