+++ test/user/UserTest.php
@@ -1,5 +1,5 @@
 <?php
-use R794021\User\{AbstractUser, User};
+use R794021\User\{User};
 use R794021\Exception\DataDomainException;
 
 const USER_HAS_NO_ID_FIELD_EXCEPTION_TEXT = 'User id field should exist';
@@ -10,7 +10,7 @@ const USER_ID_FIELD_NOT_POSITIVE_INTEGER_EXCEPTION_TEXT =
 // Test if 'id' field exist in user data
 try {
     $hasException = false;
-    $user = new class (USER_HAS_NO_ID_FIELD) extends AbstractUser implements User {
+    $user = new class (USER_HAS_NO_ID_FIELD) extends User {
         public static function isContractor(): bool { return false; }
         public static function isCustomer(): bool { return false; }
 
@@ -25,7 +25,7 @@ try {
 // Test if 'id' field is a positive integer
 try {
     $hasException = false;
-    $user = new class (USER_ID_FIELD_NOT_POSITIVE_INTEGER) extends AbstractUser implements User {
+    $user = new class (USER_ID_FIELD_NOT_POSITIVE_INTEGER) extends User {
         public static function isContractor(): bool { return false; }
         public static function isCustomer(): bool { return false; }
